+++ CHANGELOG.md
@@ -6,7 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.2] - 2016-10-06
+## [1.0.3] - 2019-10-06
+### Added
+- Mirai decoder
+
+### Changed
+- Switched pprint to json indent
+
+
+## [1.0.2] - 2019-10-06
 ### Changed
 - Removed print statement from saefko decoder
 - Travis badge to .org not .com